Renamed src/ to app/
I have renamed the src folder to app. This seems to work fine, but it may not be a good idea.
However, I want to think that the project is beautiful and app seems more beautiful to me than src. It is still only 3 letters which is nice, but is a full word as opposed to a unix-y abbreviation.
Might delete later…