Drizzle Reset Function
custom script to delete migrations and db from the project, then re-gen the migrations and run them. Useful in dev when you decide to change the db in a way that is easier to just start fresh. Could add a seeding script also in future.

Avaialble as `bun run db:reset` to just get rid of db (with confirmation) or `bun run db:rebuild` to first delete (with confirmation) and then generate and migrate.