Changed to use: action
Instead of a component, it is now implemented as a svelte action. This means is can be used on any input element with `use:formatInput={pattern}` and combined with other actions. This makes the functionality much more adaptable and useful.

I’ve also refactored the code to be much nicer.

It still does not deal with trailing backspaces entirely correctly. I’m halfway there to fixing it.