Redirect after login
When visiting a protected route, if you are not yet authenticated, we redirect you to login, but now store that initial route requested and bring you back there after you login.

This does not impact authorization issues. That is a separate design problem.

This uses cookies to store the redirect path so logging in across devices will not do anything differently.